Make AbstractRecordReader its own class
* Create AccumuloRecordReader from AbstractRecordReader
* Remove intermediate abstract Input classes
